pyproject.toml
[project]
name = "2025-1-aisp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas==2.2.3",
    "scikit-learn==1.6.1",
    "scrapy-playwright==0.0.43",
    "streamlit==1.43.2",
    "streamlit-option-menu>=0.4.0",
    "torch==2.6.0",
    "transformers==4.50.3",
]
